refactoring to remove dynamic
